Moved Twitter stuff to separate file
